requirements.txt
streamlit>=1.32
gspread>=6.0.0
google-auth>=2.29.0
python-dotenv>=1.0.1
python-dateutil>=2.9.0.post0
